add gatsby images in customer sucess stories carousel
