+++ src/assets/images/customers/propagad/slider/6.png
