+++ src/components/customer-success-stories/detail-carousel/Icon.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Icon = styled.img`
+  width: 100%;
+  height: auto;
+`;
+
+export const IconContainer = styled.div`
+  width: auto;
+  height: auto;
+  &:before,
+  &after {
+    display: none;
+  }
+`;
