+++ src/pages/customer-success-stories/4geeks-academy.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import { graphql } from 'gatsby';
 import { Section, Container, Columns, Column } from 'bloomer';
 import DetailLayout from '../../components/customer-success-stories/DetailLayout';
 import DetailTitle from '../../components/customer-success-stories/DetailTitle';
@@ -7,7 +9,7 @@ import DetailParagraph from '../../components/customer-success-stories/DetailPar
 import DetailSectionImageLeft from '../../components/customer-success-stories/DetailSectionImageLeft';
 import DetailSectionImageRight from '../../components/customer-success-stories/DetailSectionImageRight';
 // import DetailTeam from '../../components/customer-success-stories/DetailTeam';
-import DetailCarousel from '../../components/customer-success-stories/DetailCarousel';
+import DetailCarousel from '../../components/customer-success-stories/detail-carousel/DetailCarousel';
 import DetailVideo from '../../components/customer-success-stories/DetailVideo';
 import DetailImageLogo from '../../components/customer-success-stories/DetailImageLogo';
 import DetailHeroRightContainer from '../../components/customer-success-stories/DetailHeroRightContainer';
@@ -17,18 +19,9 @@ import { List, ListItem } from '../../components/Typography/List';
 import { TextOrange } from '../../components/Typography/TextHelpers';
 
 import logo from '../../assets/images/customers/4geek/logo.png';
-
 import textImg1 from '../../assets/images/customers/4geek/text-img-1.png';
-//slider images
-import slider1 from '../../assets/images/customers/4geek/slider/1.png';
-import slider2 from '../../assets/images/customers/4geek/slider/2.png';
-import slider3 from '../../assets/images/customers/4geek/slider/3.png';
-import slider4 from '../../assets/images/customers/4geek/slider/4.png';
-import slider5 from '../../assets/images/customers/4geek/slider/5.png';
-import slider6 from '../../assets/images/customers/4geek/slider/6.png';
-import slider7 from '../../assets/images/customers/4geek/slider/7.png';
 
-const GeeksAcademy = () => (
+const GeeksAcademy = ({ data }) => (
   <DetailLayout>
     <Section>
       <Container>
@@ -134,29 +127,7 @@ const GeeksAcademy = () => (
      </Section> */}
 
     <Section isPaddingless>
-      <DetailCarousel>
-        <div>
-          <img src={slider1} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider2} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider3} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider4} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider5} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider6} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider7} alt="jobcore app" />
-        </div>
-      </DetailCarousel>
+      <DetailCarousel data={data.slider.edges} />
     </Section>
 
     <Section isPaddingless>
@@ -164,5 +135,31 @@ const GeeksAcademy = () => (
     </Section>
   </DetailLayout>
 );
+
+GeeksAcademy.propTypes = {
+  data: PropTypes.object.isRequired,
+};
+
+export const pageQuery = graphql`
+  query {
+    slider: allFile(filter: { relativeDirectory: { eq: "customers/4geek/slider" } }) {
+      edges {
+        node {
+          id
+          name
+          childImageSharp {
+            fluid(quality: 100) {
+              src
+              srcSet
+              sizes
+              aspectRatio
+              base64
+            }
+          }
+        }
+      }
+    }
+  }
+`;
 
 export default GeeksAcademy;
