+++ src/pages/customer-success-stories/collabtogrow.js
@@ -7,8 +7,8 @@ import DetailParagraphTitle from '../../components/customer-success-stories/Deta
 import DetailParagraph from '../../components/customer-success-stories/DetailParagraph';
 import DetailSectionImageLeft from '../../components/customer-success-stories/DetailSectionImageLeft';
 import DetailSectionImageRight from '../../components/customer-success-stories/DetailSectionImageRight';
-import DetailTeam from '../../components/customer-success-stories/DetailTeam';
-import DetailCarousel from '../../components/customer-success-stories/DetailCarousel';
+// import DetailTeam from '../../components/customer-success-stories/DetailTeam';
+// import DetailCarousel from '../../components/customer-success-stories/detail-carousel/DetailCarousel';
 import DetailVideo from '../../components/customer-success-stories/DetailVideo';
 import DetailImageLogo from '../../components/customer-success-stories/DetailImageLogo';
 import DetailHeroRightContainer from '../../components/customer-success-stories/DetailHeroRightContainer';
@@ -199,15 +199,16 @@ const Collabtogrow = () => (
       </DetailParagraph>
     </DetailSectionImageLeft>
 
-    <Section>
+    {/* <Section>
       <Container>
-        <DetailTeam />
+      <DetailTeam />
       </Container>
-    </Section>
+     </Section> 
 
+  
     <Section isPaddingless>
-      <DetailCarousel />
-    </Section>
+      <DetailCarousel  />
+    </Section>*/}
 
     <Section isPaddingless>
       <DetailsOtherStories />
