+++ src/pages/customer-success-stories/jobcore.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import { graphql } from 'gatsby';
 import { Section, Container, Columns, Column } from 'bloomer';
 import DetailLayout from '../../components/customer-success-stories/DetailLayout';
 import DetailTitle from '../../components/customer-success-stories/DetailTitle';
@@ -9,7 +11,7 @@ import DetailSectionImageLeft from '../../components/customer-success-stories/De
 import DetailSectionImageRight from '../../components/customer-success-stories/DetailSectionImageRight';
 import DetailImagesDescription from '../../components/customer-success-stories/DetailImagesDescription';
 // import DetailTeam from '../../components/customer-success-stories/DetailTeam';
-import DetailCarousel from '../../components/customer-success-stories/DetailCarousel';
+import DetailCarousel from '../../components/customer-success-stories/detail-carousel/DetailCarousel';
 import DetailImageLogo from '../../components/customer-success-stories/DetailImageLogo';
 import DetailHeroRightContainer from '../../components/customer-success-stories/DetailHeroRightContainer';
 import DetailsOtherStories from '../../components/customer-success-stories/DetailsOtherStories';
@@ -21,15 +23,8 @@ import text from '../../assets/images/customers/jobcore/Texto-corto.jpg';
 
 import textImg1 from '../../assets/images/customers/jobcore/text-img-1.png';
 import textImg2 from '../../assets/images/customers/jobcore/text-img-2.jpg';
-//slider images
-import slider1 from '../../assets/images/customers/jobcore/slider/Job1.jpg';
-import slider2 from '../../assets/images/customers/jobcore/slider/Job2.jpg';
-import slider3 from '../../assets/images/customers/jobcore/slider/Job3.jpg';
-import slider4 from '../../assets/images/customers/jobcore/slider/job4.jpg';
-import slider5 from '../../assets/images/customers/jobcore/slider/job5.jpg';
-import slider6 from '../../assets/images/customers/jobcore/slider/job6.jpg';
 
-const Jobcore = () => (
+const Jobcore = ({ data }) => (
   <DetailLayout>
     <Section>
       <Container>
@@ -143,31 +138,38 @@ const Jobcore = () => (
      </Section> */}
 
     <Section isPaddingless>
-      <DetailCarousel>
-        <div>
-          <img src={slider4} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider1} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider2} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider3} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider5} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider6} alt="jobcore app" />
-        </div>
-      </DetailCarousel>
+      <DetailCarousel data={data.slider.edges} />
     </Section>
     <Section isPaddingless>
       <DetailsOtherStories />
     </Section>
   </DetailLayout>
 );
+
+Jobcore.propTypes = {
+  data: PropTypes.object.isRequired,
+};
+
+export const pageQuery = graphql`
+  query {
+    slider: allFile(filter: { relativeDirectory: { eq: "customers/jobcore/slider" } }) {
+      edges {
+        node {
+          id
+          name
+          childImageSharp {
+            fluid(quality: 100) {
+              src
+              srcSet
+              sizes
+              aspectRatio
+              base64
+            }
+          }
+        }
+      }
+    }
+  }
+`;
 
 export default Jobcore;
