+++ src/pages/customer-success-stories/massone.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import { graphql } from 'gatsby';
 import { Section, Container, Columns, Column } from 'bloomer';
 import DetailLayout from '../../components/customer-success-stories/DetailLayout';
 import DetailTitle from '../../components/customer-success-stories/DetailTitle';
@@ -6,7 +8,7 @@ import DetailSubTitle from '../../components/customer-success-stories/DetailSubT
 import DetailParagraph from '../../components/customer-success-stories/DetailParagraph';
 import DetailSectionImageLeft from '../../components/customer-success-stories/DetailSectionImageLeft';
 // import DetailTeam from '../../components/customer-success-stories/DetailTeam';
-import DetailCarousel from '../../components/customer-success-stories/DetailCarousel';
+import DetailCarousel from '../../components/customer-success-stories/detail-carousel/DetailCarousel';
 import DetailImageLogo from '../../components/customer-success-stories/DetailImageLogo';
 import DetailHeroRightContainer from '../../components/customer-success-stories/DetailHeroRightContainer';
 import DetailsOtherStories from '../../components/customer-success-stories/DetailsOtherStories';
@@ -19,19 +21,8 @@ import logo from '../../assets/images/customers/massone/logo.jpg';
 import text from '../../assets/images/customers/massone/text-corto.jpg';
 
 import textImg1 from '../../assets/images/customers/massone/text-img-1.png';
-//slider images
-import slider1 from '../../assets/images/customers/massone/slider/ma-1.jpg';
-import slider2 from '../../assets/images/customers/massone/slider/ma-2.jpg';
-import slider3 from '../../assets/images/customers/massone/slider/ma-3.jpg';
-import slider4 from '../../assets/images/customers/massone/slider/ma-4.jpg';
-import slider5 from '../../assets/images/customers/massone/slider/ma-5.jpg';
-import slider6 from '../../assets/images/customers/massone/slider/ma-6.jpg';
-import slider7 from '../../assets/images/customers/massone/slider/ma-7.jpg';
-import slider8 from '../../assets/images/customers/massone/slider/ma-8.jpg';
-import slider9 from '../../assets/images/customers/massone/slider/ma-9.jpg';
-import slider10 from '../../assets/images/customers/massone/slider/ma-10.jpg';
 
-const Massone = () => (
+const Massone = ({ data }) => (
   <DetailLayout>
     <Section>
       <Container>
@@ -110,38 +101,7 @@ const Massone = () => (
      </Section> */}
 
     <Section isPaddingless>
-      <DetailCarousel>
-        <div>
-          <img src={slider1} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider2} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider3} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider4} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider5} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider6} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider7} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider8} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider9} alt="jobcore app" />
-        </div>
-        <div>
-          <img src={slider10} alt="jobcore app" />
-        </div>
-      </DetailCarousel>
+      <DetailCarousel data={data.slider.edges} />
     </Section>
 
     <Section isPaddingless>
@@ -149,5 +109,31 @@ const Massone = () => (
     </Section>
   </DetailLayout>
 );
+
+Massone.propTypes = {
+  data: PropTypes.object.isRequired,
+};
+
+export const pageQuery = graphql`
+  query {
+    slider: allFile(filter: { relativeDirectory: { eq: "customers/massone/slider" } }) {
+      edges {
+        node {
+          id
+          name
+          childImageSharp {
+            fluid(quality: 100) {
+              src
+              srcSet
+              sizes
+              aspectRatio
+              base64
+            }
+          }
+        }
+      }
+    }
+  }
+`;
 
 export default Massone;
